+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.14.0] - 2022-02-15
+### Added
+- Video ads support
+### Changed
+- 2FA requirement when setting password
+- 2FA requirement when connecting the wallet
+
 ## [1.13.2] - 2022-02-10
 ### Changed
 - Token on referral links list is link
@@ -361,7 +368,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Publisher features (Sites & AdUnits)
 
  
-[Unreleased]: https://github.com/adshares/adpanel/compare/v1.13.2...develop
+[Unreleased]: https://github.com/adshares/adpanel/compare/v1.14.0...develop
+[1.14.0]: https://github.com/adshares/adpanel/compare/v1.13.2...v1.14.0
 [1.13.2]: https://github.com/adshares/adpanel/compare/v1.12.0...v1.13.2
 [1.12.0]: https://github.com/adshares/adpanel/compare/v1.11.2...v1.12.0
 [1.11.2]: https://github.com/adshares/adpanel/compare/v1.11.1...v1.11.2